refactor: move all dummies to testhelper directory
