+++ internal/repository/seller/seller_find_by_id_test.go
@@ -10,6 +10,7 @@ import (
 
 	repository "github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/repository/seller"
 	models "github.com/varobledo_meli/W17-G10-Bootcamp.git/pkg/models/seller"
+	testhelpers "github.com/varobledo_meli/W17-G10-Bootcamp.git/testhelpers"
 )
 
 func TestSellerRepository_FindById(t *testing.T) {
@@ -27,7 +28,7 @@ func TestSellerRepository_FindById(t *testing.T) {
 		{
 			name: "success",
 			mock: func(mock sqlmock.Sqlmock, id int) {
-				s := repository.SellersMapStub[1]
+				s := testhelpers.SellersMapStub[1]
 				rows := sqlmock.NewRows([]string{"id", "cid", "company_name", "address", "telephone", "locality_id"}).
 					AddRow(s.Id, s.Cid, s.CompanyName, s.Address, s.Telephone, s.LocalityId)
 				mock.ExpectQuery("^SELECT (.+) FROM sellers WHERE id").
@@ -37,7 +38,7 @@ func TestSellerRepository_FindById(t *testing.T) {
 			args:    args{id: 1},
 			wantErr: false,
 			expectedResult: func() *models.Seller {
-				s := repository.SellersMapStub[1]
+				s := testhelpers.SellersMapStub[1]
 				return &s
 			}(),
 		},
