+++ internal/repository/seller/seller_update_test.go
@@ -11,6 +11,7 @@ import (
 
 	repository "github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/repository/seller"
 	models "github.com/varobledo_meli/W17-G10-Bootcamp.git/pkg/models/seller"
+	testhelpers "github.com/varobledo_meli/W17-G10-Bootcamp.git/testhelpers"
 )
 
 func TestSellerRepository_Update(t *testing.T) {
@@ -33,7 +34,7 @@ func TestSellerRepository_Update(t *testing.T) {
 					WithArgs(s.Cid, s.CompanyName, s.Address, s.Telephone, s.LocalityId, s.Id).
 					WillReturnResult(sqlmock.NewResult(0, 1))
 			},
-			args:    args{id: 1, seller: repository.SellersMapStub[1]},
+			args:    args{id: 1, seller: testhelpers.SellersMapStub[1]},
 			wantErr: false,
 		},
 		{
@@ -47,7 +48,7 @@ func TestSellerRepository_Update(t *testing.T) {
 					WithArgs(s.Cid, s.CompanyName, s.Address, s.Telephone, s.LocalityId, s.Id).
 					WillReturnError(mysqlErr)
 			},
-			args:           args{id: 1, seller: repository.SellersMapStub[1]},
+			args:           args{id: 1, seller: testhelpers.SellersMapStub[1]},
 			wantErr:        true,
 			expectedErrMsg: "cid is already used",
 		},
@@ -62,7 +63,7 @@ func TestSellerRepository_Update(t *testing.T) {
 					WithArgs(s.Cid, s.CompanyName, s.Address, s.Telephone, s.LocalityId, s.Id).
 					WillReturnError(mysqlErr)
 			},
-			args:           args{id: 1, seller: repository.SellersMapStub[1]},
+			args:           args{id: 1, seller: testhelpers.SellersMapStub[1]},
 			wantErr:        true,
 			expectedErrMsg: "locality_id is already used",
 		},
@@ -77,7 +78,7 @@ func TestSellerRepository_Update(t *testing.T) {
 					WithArgs(s.Cid, s.CompanyName, s.Address, s.Telephone, s.LocalityId, s.Id).
 					WillReturnError(mysqlErr)
 			},
-			args:           args{id: 1, seller: repository.SellersMapStub[1]},
+			args:           args{id: 1, seller: testhelpers.SellersMapStub[1]},
 			wantErr:        true,
 			expectedErrMsg: "data conflict",
 		},
@@ -92,7 +93,7 @@ func TestSellerRepository_Update(t *testing.T) {
 					WithArgs(s.Cid, s.CompanyName, s.Address, s.Telephone, s.LocalityId, s.Id).
 					WillReturnError(mysqlErr)
 			},
-			args:           args{id: 1, seller: repository.SellersMapStub[1]},
+			args:           args{id: 1, seller: testhelpers.SellersMapStub[1]},
 			wantErr:        true,
 			expectedErrMsg: "locality does not exist",
 		},
@@ -103,7 +104,7 @@ func TestSellerRepository_Update(t *testing.T) {
 					WithArgs(s.Cid, s.CompanyName, s.Address, s.Telephone, s.LocalityId, s.Id).
 					WillReturnError(errors.New("connection lost"))
 			},
-			args:           args{id: 1, seller: repository.SellersMapStub[1]},
+			args:           args{id: 1, seller: testhelpers.SellersMapStub[1]},
 			wantErr:        true,
 			expectedErrMsg: "internal server error",
 		},
